feat: add initial files and options and create WebSocket server
